Improved output file size calculation with an explicit cast from off_t to size_t for the file length

+++ src/pcat.cxx
@@ -95,7 +95,7 @@ namespace pcat
 		constexpr std::size_t zero{};
 		return std::accumulate(inputFiles.begin(), inputFiles.end(), zero,
 			[](const std::size_t count, const fd_t &file) noexcept -> std::size_t
-				{ return count + file.length(); }
+				{ return count + static_cast<std::size_t>(file.length()); }
 		);
 	}
